Refactor ShopCard and ShopItem props.
Replaced individual 'name', 'price', 'color', and 'img' props in ShopCard and ShopItem with a single 'card' or 'item' prop. This allows for easier and more organized passing of product data to these components. Adjustments were made accordingly in dependent 'ListView' and 'CardsView' components to pass whole item/card instead of individual fields. This makes the code more modular and easier to read and maintain.