Update layout and stylings of shopping components
A significant overhaul of the layout and stylings for the 'ShopCard' and 'ShopItem' components was carried out. The changes aim to improve readability and aesthetics in the application. This includes positioning items in a more organized manner, adjusting font sizes, colors and using flexbox properties for a cleaner layout. 'IconSwitch', 'ListView' and 'CardsView' components were also adjusted accordingly. This change provides an improved user experience with a more visually appealing design. Lastly, changes in the 'page.tsx' were made to accommodate the revised layout.